testing

+++ pid.h
@@ -4,12 +4,12 @@
 #include <stdlib.h>
 #include <time.h>
 
-extern short Kp;
-extern short Ki;
-extern short Kd;
-extern short pe; // proportional error
-extern short ie; // integral error
-extern short de; // derivative error
+extern int Kp;
+extern int Ki;
+extern int Kd;
+extern float pe; // proportional error
+extern float ie; // integral error
+extern float de; // derivative error
 extern float set_point;
 extern float error;
 
@@ -18,8 +18,8 @@ void pid_controller_loop(void);
 
 //PID data processing subroutines
 float derivative(void);
-short integral(void);
-void circshift_single_insert(int* array, int element, int size);
+int integral(void);
+void circshift_single_insert(float* array, int element, int size);
 
 
 #endif
